Acl Done

+++ app/Http/Controllers/API/UserController.php
@@ -19,7 +19,12 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+       // $this->authorize('isAdmin');
+        if(\Gate::allows('isAdmin') || \Gate::allow('isAuthor')){
+            return User::latest()->paginate(10); 
+        }
+        
+       
     }
 
     /**
@@ -33,7 +38,7 @@ class UserController extends Controller
         $this->validate($request,[
             'name'=> 'required|string|max:119',
             'email'=> 'required|email|unique:users',
-            'password'=> 'required|min:6',
+            'password' => 'required',
             'type'=> 'required',
             'photo'=> '',
             'bio'=> 'required'
@@ -66,11 +71,30 @@ class UserController extends Controller
 
     public function updateProfile(Request $request)
     {
+       
         $user = auth('api')->user();
-        // return ['message'=>"success"];
+         $this->validate($request,[
+            'name' => 'required|string|max:191',
+            'email' => 'required|string|email|max:191|unique:users,email,'.$user->id,
+            'password' => 'sometimes|required|min:6'
+        ]);
+
+        $currentPhoto=$user->photo;
+       if($request->photo != $currentPhoto){
         $name = time().'.' . explode('/', explode(':', substr($request->photo, 0, strpos($request->photo, ';')))[1])[1];
             \Image::make($request->photo)->save(public_path('img/profile/').$name);
             $request->merge(['photo' => $name]);
+             $userPhoto = public_path('img/profile/').$currentPhoto;
+            if(file_exists($userPhoto)){
+                @unlink($userPhoto);
+            }
+           }
+
+        if(!empty($request->password)){
+            $request->merge(['password' => Hash::make($request['password'])]);
+        }
+
+            $user->update($request->all());
     }
 
     /**
@@ -101,7 +125,8 @@ class UserController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
+    {   
+        $this->authorize('isAdmin');
         $user = User::findOrFail($id);
         $user->delete();
     }
